Fix compilation under VS2015 Fix using of 'uint' vertex attribute Add a new swig_interface_test for later use with SWIG Remove unused earth_plates_test

+++ assets/shaders/gaussian_splatting.vert.glsl
@@ -23,7 +23,7 @@ struct ShcoefData
 layout(std140, binding = 4) restrict readonly buffer ShcoefBuffer { ShcoefData shcoef[]; };
 #  endif
 
-VERSE_VS_IN float osg_UserIndex;
+VERSE_VS_IN uint osg_UserIndex;
 VERSE_VS_OUT vec4 color, invCovariance;
 VERSE_VS_OUT vec2 center2D;
